osd/PGLog: delay populating PG log hashmaps
When booting up OSD, it loads all PGs and their respective logs. To speed
up processing later, these logs are accompanied by separate unordered_maps
which are also populated during PG load.
Delay that until we actually need to access it, so we don't occupy too much
memory right from start - and when we need it, populate just the map that
we want to use, not all 3 of them at once.

Signed-off-by: Piotr Dałek <piotr.dalek@ts.fujitsu.com>