librados: add map_object API
Ceph use crush algorithm to provide the mapping of objects to OSD servers.
However there's public API for this instead you'll have to rely on the
command to monitor:

$ ceph osd map pool_name obj_name

The "ceph_get_osd_crush_location" method is made available via libcephfs
but not avaiale in librados.

This patch provides API to render the object distribution layout. With this
the application could choose which nodes to access, for load-balance purpose.

Signed-off-by: Yuan Zhou <yuan.zhou@intel.com>