rgw: keystone v3 support
We've improved the current rgw to support authv3 using keystone.

Rather than introduce a factory and inheritance of the keystone token structure, the token now accommodates both formats and the parser populates the relevant structure dependent on the version of the API and then normalises it to the v3 layout, which uses project rather than tenant.

There was also some big code moves which were required to allow us to improve the separation of concerns regarding keystone interaction from within the swift class and required the cms code to be extracted so it could be included in different places.

As for documentation there are new options in ceph.conf

rgw keystone api version = [ 2.0 / 3 ] ( default = 2.0 )

For admin authentication without a token, the following is required

rgw keystone admin domain = acme.org
rgw keystone admin project = project_name

Signed-off-by: Mark Barnes <mark.barnes@ocado.com>