osd: only calculate op crush mapping if we don't have the PG
We were pessimistically calculating the request mapping to ensure
that it is really meant for us before looking up the PG.  Instead,
look up the PG, and queue it there if we have it--the PG already
does the necessary checks in PG::can_discard_op().  If we
don't have the PG, *then* do the crush mapping and extra checks.

Reported-by: Somnath Roy <somnath.roy@sandisk.com>
Signed-off-by: Sage Weil <sage@redhat.com>