suite: add the ceph-qa-suite hash to the job config
To be able to re-run a job we need:

* the ceph repository and hash
* the ceph-qa-suite repository and hash
* the job description

When reading the job log from the archive directory, we currently only
have the ceph-qa-suite branch name. If and when the branch name moves
on, the job description may cease to be valid and re-running the job
may no longer be possible.

Signed-off-by: Loic Dachary <ldachary@redhat.com>