os/FileStore: For clone, first clone and later trucated size.
When do clone(copy obA to obB). It firstly truncated size of obB to
zero. Later do clone(data and omap).
If we firstly do clone and later trucated size of obB to size of obA.
This will reduce truncate overhead when a)obB don't exist b)size obB >
obA.

Signed-off-by: Jianpeng Ma <jianpeng.ma@intel.com>