Website feature
As same as amazon S3 interface, the website attribute of bucket have
been added. The "ErrorDocument" and "redirect rules" of website have not
been realized now. The "IndexDocument" of website have been implemented.

To configure a bucket as a website, you can add this subresource on the
bucket with website configuration information such as the file name of
the index document and any redirect rules.

For example:
s3cmd ws-create s3://mytest --ws-index=index.html
curl http://localhost/mytest  -L
the return of curl will be the content of index.html

Signed-off-by: sunfangchen <richard.sun1999@gmail.com>