OSD: Option to delete transaction object from FileStore worker thread
Presently, the transaction object is been deleted by the Finisher thread
and if we unleash the journal throttle, memory is piling up even if we
throttle the filestore op_wq to a certain bytes. Based on a flag
provided by caller the transaction object is now deleted from the
FileStore worker thread itself right after transaction execution.
We can increase the journal commit rate because of this and thus helping
to improve write performance.

Signed-off-by: Somnath Roy <somnath.roy@sandisk.com>