WBThrottle: fix incorrect throttle
This bug is introduced in commit edf9505, that throttle()
incorrectly use the "start flusher" throttle value to block new
write requests.

On a test environment of 2 storage nodes with 7 HDD each, the
overall write performance measured by iozone is doubled with this
patch (boosted from 160MB/s to 300MB/s).

Signed-off-by: Zhang Huan <zhanghuan@ict.ac.cn>