common/buffer: changed buffer from class to namespace
Since the buffer only contained static functions it doesn't make sense
for it to be a class. Changing it to a namespace will allow
bufferlist, bufferptr and bufferhash to be forward declared.
This wasn't possible when they were nested classes.

Function declarations for
inc_total_alloc and
dec_total_alloc
were removed from the header file and changed to static, to simulate
the private access specifier, which they were under.

Moved the operators inside the buffer namespace to take advantage of
the Argument Dependent Lookup.

Moved typedefs to  buffer_fwd header file which we can include when
only the forward declarations are needed.

Since the buffer is not a class anymore all the instances of
using ceph::buffer
were removed.

Signed-off-by: Michal Jarzabek <stiopa@gmail.com>