mon/OSDMonitor: block 'ceph osd pg-temp ...' if update is pending
The OSD expects it's pg_temp update requests to succeed.  If it
races with an ill-timed admin request, it can get stuck in
WaitActingChange indefinitely.

This is only a real problem now that the OSD/mon interaction has
been updated with wip-bigbang; previously we would retry (although
it would take a while).  Backporting is optional.

Signed-off-by: Sage Weil <sage@redhat.com>