ceph.spec.in: use %tmpfiles_create macro
Ensure the macro is defined, and use it to placate the relevant openSUSE
RPMLINT check.

Drop --prefix=/run/ceph (introduced by ad4cf684) because we now supply the
configuration file on the systemd-tmpfiles command line. (We don't need to run
systemd-tmpfiles on all files in /usr/lib/tmpfiles.d/ - just on ours.)

Signed-off-by: Nathan Cutler <ncutler@suse.com>