openstack: TeuthologyOpenStack.cloud_init_wait continuous output
Instead of tailing the files every 30 seconds, wait for the
string that proves the teuthology installation completed with a tail -f
and display the result continously.

The number of attempts is reduced from 100 to 30 because there is no
need for repeated calls to tail to check for the desired string.

Signed-off-by: Loic Dachary <loic@dachary.org>