Revert "Revert "ceph.spec.: add epoch""
Re-add the Epoch bump to the packaging on the Firefly branch.

Ceph Hammer and newer already have this change. It's time to add it to
Firefly as well. The reasons that I removed it back in December 2014 are
no longer valid, and this will fix some interactions with the Ceph
client packages that ship in Base RHEL 7.1+.

Fixes: #11104 http://tracker.ceph.com/issues/11104
Fixes: #13794 http://tracker.ceph.com/issues/13794

This reverts commit 7faae891aefa4c21c50430fa03d9204a86d082f8.

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>