osd: make block device fsid probing generic
Currently the option name and invocation assume that the block device
is a journal (and FileStore journal, managed by FileJournal).  Rework
the interface so that we can probe any block device and other ObjectStore
implementations will have a chance to identify the device (and return the
osd fsid).

Switch to a static method while we are at it so we avoid instantiating
each backend.

Note that only FileStore is probed at the moment; that will change soon!

Signed-off-by: Sage Weil <sage@redhat.com>