osd: fix ScrubJob compare function
If we not consider the deadline in compare function. Some pgs would not
be scheduled when the deadline arrives.

Such as, there is ScrubJob A in sched_scrub_pg. A.pgid = 1.1, A.sched_time = 0:00,
A.deadline = 23:00. Now we send a command `ceph pg scrub 1.2`, so osd would
add a new ScrubJob B. It is possible that B.pgid = 1.2, A.sched_time = 0:00,
A.deadline = 0:00. Current strategy ScrubJob B would be scheduled after A.
But A.deadline is much later than B.deadline. So it possible that job B is
not scheduled when the job B deadline arrives.

Signed-off-by: Xinze Chi <xinze@xsky.com>