FileJournal: New journal padding mode
Current, because the requirement of DirectIo, the padding mode
is adding padding data for every entry.

In fact, we fetch a couple of entries as a system write as possible
as can. So we only make sure system-write content met the requirement
of DirectIO. We adding a padding entry at the end of system-write
content(for most case, this include a couple of entries).

User could set the osd_journal_padding_ahead to decide which mode.
osd_journal_padding_ahead = true is good choice for pcie-ssd.

Signed-off-by: Jianpeng Ma <jianpeng.ma@intel.com>
Signed-off-by: Xinze Chi <xinze@xsky.com>