tests: control the scrub min interval when necessary
When forcing scrub or repair for test purposes, the ScrubJob may be
created right before the scrub or repair command is sent to the OSD.  It
will set its scheduled time to something a function of
osd_scrub_min_interval and osd_scrub_interval_randomize_ratio. It will
default to more than 24h and the test will timeout before it can happen.

The osd_scrub_min_interval is set to one second when such tests are
done. It can't be set for all tests otherwise it will make all OSDs
scrub constantly and most probably break things.

http://tracker.ceph.com/issues/13986 Fixes: #13986

Signed-off-by: Loic Dachary <loic@dachary.org>