common/buffer: reset iterator on rebuild
When we rebuild we implicitly invalidate any iterator into the
old list of bufferptrs.  Reset the internal last_p iterator to
fix.

The test to reproduce the bug is tricky: we need to get into
a situation where last_p is initialized and p points to
_buffers.begin() (not .end()), e.g.,

  bufferlist bl, bl2;
  bl2.append("bar");
  bl.swap(bl2);

and then rebuild the buffer, so that none of the checks in the
copy_out path reset last_p for some other reason.

Signed-off-by: Sage Weil <sage@redhat.com>