osd/ReplicatedPG: check obs.exists for  CEPH_OSD_OP_CMPXATTR/SRC_CMPXATTR.
We may use those ops(CMPXATTR, WRITE, SET_XATTR) to do a write. Using
CMPXATTR to indiate whether to write. But if object don't exist, because
the ops is writeable(because WRITE, SET_XATTR), so it can do CMPXATTR.
But at this situation, the func getattr_maybe_cache return -ENOENT.
For optimization, we first check obs.exists.

Signed-off-by: Jianpeng Ma <jianpeng.ma@intel.com>