Dispatcher.h: include assert.h
The inline ms_fast_dispatch implementation calls assert, which
puts asset.h in the interface.

The fact that many Dispatcher descendants don't implement
ms_fast_dispatch prevents making the method pure virtual--which
suggests that maybe there is a need for a FastDispatcher
interface that inherits Dispatcher and introduces ms_fast_dispatch
and related?

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>