osd: do not keep ref of old osdmap in pg
do not hold a strong reference to last_persisted_osdmap in PG. as an
OSD tries to trim previously persisted osdmaps, if they are not
referenced anymore. this helps to keep the meta collection in a
manageable size. if we advance the osdmap many times, and some PGs
are not impacted by these changes, it's very likely that they are
still holding very old osdmap references in their `last_persisted_osdmap_ref`.
this practically prevents the OSD from removing the out-dated osdmap
in OSD::handle_osd_map() if `last_persisted_osdmap_ref` is not updated
in a timely manner, for example, due to a large
"osd_pg_epoch_persisted_max_stale". so, instead of holding a reference
of last_persisted_osdmap, we can simply remember its epoch.

See also: #13990

Signed-off-by: Kefu Chai <kchai@redhat.com>