ceph-disk: deactivate must unmap dmcrypted journal
If the journal is not unmapped, ceph-disk destroy will fail to zap the
corresponding devices because it is still held by devicemapper.

A consequence of this modification is that

   ceph-disk activate --dmcrypt --reactivate

no longer works from the command line, because it does not map the
dmcrypted journal. The --reactivate option is added to activate-journal
which will map both the journal and the data devices, if necessary.

http://tracker.ceph.com/issues/14233 Fixes: #14233

Signed-off-by: Loic Dachary <loic@dachary.org>