ceph-disk: list accepts absolute dev names
The ceph-disk list subcommand now accepts /dev/sda as well as sda.
The filtering is done on the full list of devices instead of restricting
the number of devices explored. Always obtaining the full list of
devices makes things simpler when trying to match a dmcrypted device to
the corresponding raw device.

Signed-off-by: Loic Dachary <loic@dachary.org>