README: fix rst syntax for sub-bullets
The install task contains its own list of sub-point bullets.  Prior to
this change, rST did not treat this list as a sub-list of the "install
task" bullet point. Instead, it rendered them at the same level as the
install task and all the other tasks.

Indent the sub-bullets under the install task's description. This will
cause rST to treat them as true sub-bullets under the install task.