mon: Monitor: get rid of weighted clock skew reports
By weighting the reports we were making it really hard to get rid of a
clock skew warning once the cause had been fixed.

Instead, as soon as we get a clean bill of health, let's run a new round
and soon as possible and ascertain whether that was a transient fix or
for realsies. That should be better than the alternative of waiting for
an hour or something (for a large enough skew) for the warning to go
away - and with it, the admin's sanity ("WHAT AM I DOING WRONG???").

Fixes: #14175

Signed-off-by: Joao Eduardo Luis <joao@suse.de>

(cherry pick from commit 17d8ff429c7dca8fc1ada6e7cc8a7c4924a22e28)