FreeBSD return the extattr list into always in guaranteed the same order.
The attributes are not returned in the originally submitted order.
It is not guaranteed in the API spec. But this is also the case for the
Linux API. So event there this test might start failing.

As such test sometimes fails, sometimes succeeds. And is excluded in the
FreeBSD case

Signed-off-by: Willem Jan Withagen <wjw@digiware.nl>