os/bluestore/BlueStore: Simplifying default return value
r variable wasn't initialized but set to 0 twice. Once just before an early
"goto out" code and the other before the main code of the function.

This patch simply initialize r to 0 at the begining.
This make the code easier to read and prevent any future call to "out" label without a proper value stored in r.

Signed-off-by: Erwan Velu <erwan@redhat.com>