os/filestore/FileJournal: Removing dead code
Despite read_pos & last_committed_seq, the seq variable is local and created a
few lines before.

While the reinitiliasation of read_pos & last_commited_seq is useless, the seq
variable is destroyed by the return call.

That makes the "seq = 0" being useless and a dead code. This patch simply
removes it.

Signed-off-by: Erwan Velu <erwan@redhat.com>