osd: Add runtime config option to select which queue to use and the priority of the cutoff between the strict queue and the normal queue.
osd_op_queue takes prio, wpq, or debug_random to select PrioritizedQueue,
WeightedPriorityQueue or randomly select one for testing.
Default is PrioritizedQueue.

osd_op_queue_cut_off takes low, high, or debug_random to select ops with
priority greater than or equal to 64 (low) or 196 (high) to be queued in
the strict queue rather than the normal queue. It is advantageous to use
the high option so that repops are queued proportionally with regular ops
to help prevent I/O starvation when an OSD has many non-primary PGs. This
allows clients who accesses a primary PG on this busy OSD to not be
indefinatly blocked by OSD sending all repops to the front of the line.
Default is low.

The defaults preserve the original operation.

Signed-Off-By: Robert LeBlanc <robert.leblanc@endurance.com>