common: add --extra-config that defaults to $data_dir/config
This will be parsed *after* the normal config search list is
considered and (if a match is found) a config file is parsed.
The extra config file options will overlay on top of/override the
main config file options.  It is also possible to skip the main
config file entirely and use only an extra_config file.

Signed-off-by: Sage Weil <sage@redhat.com>