Changes requested for upstream patch approval.
Keystone version number an integer, as we only care about v2 to v3 differences here.

Clear commenting on end 'bracing' for json generation code in rgw_keystone.cc

We believe this makes the code much easier to maintain and follow than some abstract templating system would do.

Commenting where the end of each part of the structed json is makes following what is going on in the slightly
more complex Keystone v3 case easier to follow and therefore easier to maintain which was the request.

Signed-off-by: Martin Houston <martin.houston@ocado.com>